+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # dartlang plugin changelog
 
+## unreleased
+- added an option to change the break on exceptions mode in the debugger
+  (break on all exceptions, break on uncaught exceptions, or don't break on exceptions)
+- reduced the directory depth we look through for Dart projects
+
 ## 0.6.5
 - Updated code completion to reduce inadvertent completions after array braces
 - Dart code completion now supports fuzzy matching